Drop some (size_t) casts
Upcasts of the same type are automatic, so no need for an explicit cast.

+++ src/shared/format-table.h
@@ -111,9 +111,9 @@ int table_set_empty_string(Table *t, const char *empty);
 int table_set_display_all(Table *t);
 #if 0 /// UNNEEDED by elogind
 int table_set_display_internal(Table *t, size_t first_column, ...);
-#define table_set_display(...) table_set_display_internal(__VA_ARGS__, (size_t) SIZE_MAX)
+#define table_set_display(...) table_set_display_internal(__VA_ARGS__, SIZE_MAX)
 int table_set_sort_internal(Table *t, size_t first_column, ...);
-#define table_set_sort(...) table_set_sort_internal(__VA_ARGS__, (size_t) SIZE_MAX)
+#define table_set_sort(...) table_set_sort_internal(__VA_ARGS__, SIZE_MAX)
 int table_set_reverse(Table *t, size_t column, bool b);
 #endif // 0
 int table_hide_column_from_display(Table *t, size_t column);
